Add stats and expand other serializers
Create a stats feature, which enables adding stats of specific players in specific games. Include SerializerMethodFields to calculate fields, which can be calculated using model data. Add a link to boxscores (all of game's player stats) to games and add a link to all of specific players' stats in PlayerSerializer. Route urls and nested urls to aforementioned features. Resolves #40.